Refactor CLI to make it testable using `testscript`
