+++ integration/script/testdata/bogus.txtar
@@ -1,0 +1,2 @@
+! exec step bogus
+stderr 'No help topic for'
