+++ integration/script/testdata/crypto/help.txtar
@@ -1,0 +1,2 @@
+exec step help crypto
+stdout 'cryptographic primitives that balances completeness and safety'
