+++ integration/script/testdata/help.txtar
@@ -1,0 +1,2 @@
+exec step --help
+stdout 'plumbing for distributed systems'
